+++ workers/tests/integrationtests/test_indicator_ghs_pop_comparison_buildings.py
@@ -2,8 +2,6 @@ import asyncio
 import unittest
 from datetime import datetime
 
-from asyncpg import Record
-
 from ohsome_quality_analyst.geodatabase import client as db_client
 from ohsome_quality_analyst.indicators.ghs_pop_comparison_buildings.indicator import (
     GhsPopComparisonBuildings,
@@ -42,11 +40,6 @@ class TestIndicatorGhsPopComparisonBuildings(unittest.TestCase):
 
         self.indicator.create_figure()
         self.assertIsNotNone(self.indicator.result.svg)
-
-    @oqt_vcr.use_cassette()
-    def test_get_zonal_stats_population(self):
-        result = asyncio.run(self.indicator.get_zonal_stats_population())
-        self.assertIsInstance(result, Record)
 
 
 if __name__ == "__main__":